requirements.txt
gdown
torch
torchvision
numpy
tqdm
pillow